Reorganise code into src/ and tests/ directories
Use find_packages() in setup.py.

Drop Django 1.4 and 1.5 from automated tests.
