+++ .travis.yml
@@ -3,12 +3,9 @@ python: 2.7
 install:
   - pip install tox
 env:
-  - TOXENV=py27-django1.4
-  - TOXENV=py27-django1.5
   - TOXENV=py27-django1.6
   - TOXENV=py27-django1.7
   - TOXENV=py27-django1.8
-  - TOXENV=py34-django1.5
   - TOXENV=py34-django1.6
   - TOXENV=py34-django1.7
   - TOXENV=py34-django1.8
