+++ src/voting/managers.py
