+++ src/voting/templatetags/voting_tags.py
