+++ src/voting/views.py
