+++ tests/test_settings.py
@@ -20,5 +20,5 @@ INSTALLED_APPS = (
     'django.contrib.auth',
     'django.contrib.contenttypes',
     'voting',
-    'voting.tests',
+    'test_app',
 )
