+++ tox.ini
@@ -1,29 +1,25 @@
 [tox]
 envlist =
-    # Django 1.4 does not support Python 3
-    py27-django1.4
     # TODO: py35, django1.9
-    py{27,34}-django1.{5,6,7,8}
+    py{27,34}-django1.{6,7,8}
     coverage
 
 [testenv]
 deps =
-    django1.4: Django ~=1.4.2
-    django1.5: Django ~=1.5.0
     django1.6: Django ~=1.6.0
     django1.7: Django ~=1.7.0
     django1.8: Django ~=1.8.0
     django1.9: Django ~=1.9.0
 setenv =
-    DJANGO_SETTINGS_MODULE = voting.tests.settings
-    PYTHONPATH = {toxinidir}
+    DJANGO_SETTINGS_MODULE = test_settings
+    PYTHONPATH = {toxinidir}/tests
 commands =
-    django-admin.py test voting
+    django-admin.py test tests
 
 [testenv:coverage]
 basepython = python2.7
 commands =
-    coverage run --branch --omit=.tox/*,*/tests/*.py,*/migrations/*.py {envbindir}/django-admin.py test voting
+    coverage run --branch --omit=.tox/*,*/tests/*.py,*/migrations/*.py {envbindir}/django-admin.py test tests
 deps =
     coverage>=3.7,
     Django>=1.6,<1.7
