--- voting/tests/__init__.py
@@ -1,3 +1,0 @@
-import django
-if django.VERSION[0] == 1 and django.VERSION[1] < 6:
-    from .tests import *
